fix(components): rename detail section props file
