+++ src/components/DetailSection/detailsection.d.ts
