+++ src/components/DetailSection/index.ts
@@ -1,5 +1,5 @@
 import DetailSection from "./DetailSection";
-import type { DetailSectionProps } from "./cardetailsection";
+import type { DetailSectionProps } from "./detailsection";
 
 export type { DetailSectionProps };
 export default DetailSection;
